Better raw specs
