+++ testool/src/statetest/mod.rs
@@ -1,11 +1,13 @@
 mod executor;
 mod json;
 mod results;
+pub mod spec;
 mod suite;
 mod yaml;
 
-pub use executor::{CircuitsConfig, StateTest, StateTestError};
+pub use executor::{geth_trace, run_test, CircuitsConfig, StateTestError};
 pub use json::JsonStateTestBuilder;
 pub use results::Results;
+pub use spec::{AccountMatch, Env, StateTest, StateTestResult};
 pub use suite::{load_statetests_suite, run_statetests_suite};
 pub use yaml::YamlStateTestBuilder;
